I need to use remote_theme instead of theme: it seems.
From the docs: https://help.github.com/articles/adding-a-jekyll-theme-to-your-github-pages-site/
Activate the theme by adding a new line to your _config.yml with the theme name:
Add theme to config fileTo activate one of the officially supported themes, type theme: followed by the name of the theme (as shown in the README in the theme's source repository).
Add remote theme to config fileTo activate any other open source Jekyll theme hosted on GitHub, type remote_theme: followed by the name of the theme (as shown in the README or other documentation in the theme's source repository).